tests/wpt/tests/html/browsers/browsing-the-web/navigating-across-documents/javascript-url-global-scope.html

<!doctype html>
<meta charset=utf-8>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<a id="javascript-link" href="javascript:changeStatus()">link</a>

<script>
function changeStatus() {
  t.done();
}

var t = async_test(function(t) {
  document.querySelector("#javascript-link").click();
}, "javascript: scheme urls should be executed in current global scope");
</script>
