LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/javascript-url-global-scope-expected.txt

CONSOLE MESSAGE: line 10: TypeError: undefined is not an object (evaluating 't.done')
link

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 't.done')

TIMEOUT javascript: scheme urls should be executed in current global scope Test timed out
